test: split additive schema migration coverage
