+++ src/infra/state-migrations.media-persistence.test.ts
@@ -504,57 +504,6 @@ describe("legacy media persistence doctor migration", () => {
     }
   });
 
-  it("repairs same-version additive session schema before media validation", () => {
-    const stateDir = makeTempDir(tempDirs, "media-persistence-current-additive-");
-    const env = { OPENCLAW_STATE_DIR: stateDir };
-    const opened = openOpenClawAgentDatabase({ agentId: "main", env });
-    const databasePath = opened.path;
-    opened.db
-      .prepare(
-        `INSERT INTO session_nodes (session_key, current_session_id, entry_json, updated_at)
-         VALUES (?, ?, ?, ?)`,
-      )
-      .run(
-        "agent:main:session-1",
-        "session-1",
-        JSON.stringify({ sessionId: "session-1", updatedAt: 1 }),
-        1,
-      );
-    closeOpenClawAgentDatabasesForTest();
-    closeOpenClawStateDatabaseForTest();
-
-    const { DatabaseSync } = requireNodeSqlite();
-    const database = new DatabaseSync(databasePath);
-    database.exec(`
-      DROP TRIGGER session_nodes_entry_valid_after_insert;
-      DROP TRIGGER session_nodes_entry_valid_after_entry_update;
-      DROP TRIGGER session_nodes_entry_valid_after_identity_update;
-      DROP INDEX idx_agent_session_nodes_entry_valid_pending;
-      DROP TABLE session_key_contract;
-      ALTER TABLE session_nodes DROP COLUMN entry_valid;
-    `);
-    database.close();
-
-    const result = migrateLegacyMediaPersistence({ env });
-    expect(result.warnings).toEqual([]);
-    const repaired = new DatabaseSync(databasePath, { readOnly: true });
-    try {
-      expect(repaired.prepare("PRAGMA user_version").get()).toEqual({
-        user_version: OPENCLAW_AGENT_SCHEMA_VERSION,
-      });
-      expect(
-        repaired
-          .prepare("SELECT entry_valid FROM session_nodes WHERE session_key = ?")
-          .get("agent:main:session-1"),
-      ).toEqual({ entry_valid: 1 });
-      expect(
-        repaired.prepare("SELECT main_key FROM session_key_contract WHERE id = 1").get(),
-      ).toEqual({ main_key: "main" });
-    } finally {
-      repaired.close();
-    }
-  });
-
   it("upgrades an owned v0 database through the media prerequisite schema", () => {
     const stateDir = makeTempDir(tempDirs, "media-persistence-v0-");
     const env = { OPENCLAW_STATE_DIR: stateDir };
